config/workspace/skills/office-document-specialist-suite/requirements.txt
python-docx>=1.1.2
openpyxl
python-pptx
lxml>=5.3.0
